requirements.txt
numpy
pandas
pydeck
streamlit
streamlit_keplergl
datetime
sodapy
plotly.express